Refactor daemon internals into focused implementation modules
